requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.16.2
sqlparse==0.4.4
typing==3.7.4.3
uritemplate==4.1.1
